SnoptSolver uses C snopt_interface instead of f2c interface.
With this C interface, the accuracy of the solution changed from the
previous f2c interface. We consulted with the SNOPT author, and she
thinks we are doing the right thing. One thing to notice is that the LU
factorization at the first iteration gives different results. But both
results (f2c/snopt-interface) satisfy the default constraint/optimality
tolerance.